+++ spec/support/pre_pty.rb
@@ -2,6 +2,7 @@
 
 require 'bundler/setup'
 require_relative '../../lib/spellr/pwd'
+require_relative '../../lib/spellr/suggester'
 
 module Spellr
   module_function
@@ -12,6 +13,16 @@ module Spellr
 
   def pwd_s
     ENV['SPELLR_TEST_PWD']
+  end
+
+  class Suggester
+    class << self
+      undef slow?
+
+      def slow?
+        false
+      end
+    end
   end
 end